requirements.txt
openai
langchain
python-dotenv
sqlite3
requests
pandas
langchain-community
transformers
pysqlite3